# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ChatRequest {
  /** The user's question */
  query: string;
}

export type CitationType = (typeof CitationType)[keyof typeof CitationType];

export const CitationType = {
  document: "document",
  ticket: "ticket",
} as const;

export interface Citation {
  id: number;
  type: CitationType;
  title: string;
  url?: string;
  snippet: string;
  score: number;
}

export interface ChatResponse {
  answer: string;
  citations: Citation[];
  /** Number of retrieval steps taken (1 or 2) */
  retrieval_steps: number;
}

export interface SeedResponse {
  success: boolean;
  documents_seeded: number;
  tickets_seeded: number;
  message: string;
}

export interface SeedStatus {
  seeded: boolean;
  document_count: number;
  ticket_count: number;
}

export interface ZendeskConfig {
  api_key: string;
  subdomain: string;
}

export interface ConfluenceConfig {
  api_token: string;
  base_url: string;
  space_key?: string;
}

export interface IntegrationResponse {
  success: boolean;
  imported: number;
  message: string;
}

export interface ErrorResponse {
  error: string;
  detail?: string;
}