# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */

export interface ChatRequest {
  /** The user's question */
  query: string;
}