# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */
import type { Citation } from "./citation";

export interface ChatResponse {
  answer: string;
  citations: Citation[];
  /** Number of retrieval steps taken (1 or 2) */
  retrieval_steps: number;
}