# lib/api-zod/src/generated/types/citation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */
import type { CitationType } from "./citationType";

export interface Citation {
  id: number;
  type: CitationType;
  title: string;
  url?: string;
  snippet: string;
  score: number;
}