# lib/api-zod/src/generated/types/confluenceConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */

export interface ConfluenceConfig {
  api_token: string;
  base_url: string;
  space_key?: string;
}