# lib/api-zod/src/generated/types/seedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */

export interface SeedResponse {
  success: boolean;
  documents_seeded: number;
  tickets_seeded: number;
  message: string;
}