# lib/api-zod/src/generated/types/seedStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SupportBrainz API - RAG-powered support assistant
 * OpenAPI spec version: 0.1.0
 */

export interface SeedStatus {
  seeded: boolean;
  document_count: number;
  ticket_count: number;
}